ec/it*: Merge remaining ITE EC files
