--- src/ec/it8587e/include/ec/espi.h
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-only
-
-#ifndef _EC_ESPI_H
-#define _EC_ESPI_H
-
-// eSPI not supported on IT8587E
-#define EC_ESPI 0
-
-#endif // _EC_ESPI_H
