--- src/ec/it8587e/signature.c
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-only
-
-static __code const unsigned char __at(0x40) SIGNATURE[16] = {
-    0xA5, 0xA5, 0xA5, 0xA5, 0xA5, 0xA5, 0xA5, 0x94,
-    0x85, 0x12, 0x5A, 0x5A, 0xAA, 0x00, 0x55, 0x55,
-};
