+++ src/ec/ite/espi.c
@@ -1,7 +1,8 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
-#include <common/debug.h>
 #include <ec/espi.h>
+
+#ifdef it5570e
 
 // Not all wires are defined or implemented
 // Index 2 - AP to EC
@@ -61,3 +62,5 @@ void vw_set(struct VirtualWire * vw, enum VirtualWireState state) __critical {
     }
     *vw->index = index;
 }
+
+#endif // it5570e
