+++ src/ec/ite/include/ec/gctrl.h
@@ -8,5 +8,10 @@
 volatile uint8_t __xdata __at(0x2006) RSTS;
 volatile uint8_t __xdata __at(0x200A) BADRSEL;
 volatile uint8_t __xdata __at(0x200D) SPCTRL1;
+#ifdef it5570e
+volatile uint8_t __xdata __at(0x2030) P80H81HS;
+volatile uint8_t __xdata __at(0x2031) P80HD;
+volatile uint8_t __xdata __at(0x2032) P81HD;
+#endif
 
 #endif // _EC_GCTRL_H
