+++ src/ec/ite/include/ec/gpio.h
@@ -50,9 +50,11 @@ volatile uint8_t __xdata __at(0x16FE) GCR15;
 volatile uint8_t __xdata __at(0x16E0) GCR16;
 volatile uint8_t __xdata __at(0x16E1) GCR17;
 volatile uint8_t __xdata __at(0x16E2) GCR18;
+#ifdef it5570e
 volatile uint8_t __xdata __at(0x16E4) GCR19;
 volatile uint8_t __xdata __at(0x16E5) GCR20;
 volatile uint8_t __xdata __at(0x16E6) GCR21;
+#endif
 
 volatile uint8_t __xdata __at(0x1601) GPDRA;
 volatile uint8_t __xdata __at(0x1602) GPDRB;
