+++ src/ec/ite/signature.c
